Fixed the board crud workflow validation and the refresh-token flow

+++ client/src/services/api.js
@@ -1,4 +1,5 @@
 import axios from "axios";
+import { ChevronsDown } from "lucide-react";
 
 const api = axios.create({
   baseURL: "http://localhost:5050",
@@ -8,7 +9,7 @@ const api = axios.create({
 api.interceptors.request.use(
   (config) => {
     const token = localStorage.getItem("access-token");
-    console.log(token);
+
     if (token) {
       config.headers.Authorization = `Bearer ${token}`;
     }
@@ -21,7 +22,32 @@ api.interceptors.request.use(
 
 api.interceptors.response.use(
   (response) => response,
-  (error) => {
+  async (error) => {
+    const originalRequest = error.config;
+
+    const isRefreshRoute = originalRequest?.url?.includes(
+      "/auth/refresh-token"
+    );
+
+    if (isRefreshRoute) {
+      localStorage.clear();
+      return Promise.reject(error);
+    }
+
+    if (error.response?.status === 401 && !originalRequest._retry) {
+      originalRequest._retry = true;
+
+      const refreshResponse = await api.post("/api/v1/auth/refresh-token");
+
+      const newAccessToken = refreshResponse.data.accessToken;
+
+      localStorage.setItem("access-token", newAccessToken);
+
+      originalRequest.headers.Authorization = `Bearer ${newAccessToken}`;
+
+      return api(originalRequest);
+    }
+
     localStorage.clear();
     return Promise.reject(error);
   }
@@ -48,7 +74,7 @@ export const authService = {
     }
   },
 };
-export const taskService = {
+export const boardService = {
   createBoard: async (data) => {
     try {
       const response = await api.post("/api/v1/boards", data);
@@ -58,12 +84,8 @@ export const taskService = {
     }
   },
   getBoards: async () => {
-    try {
-      const response = await api.get("/api/v1/boards");
-      return response.data;
-    } catch (err) {
-      console.error(err);
-    }
+    const response = await api.get("/api/v1/boards");
+    return response.data;
   },
   getBoard: async (boardId) => {
     try {
@@ -73,6 +95,28 @@ export const taskService = {
       console.error(err);
     }
   },
+  createColumn: async (columns, boardId) => {
+    try {
+      const response = await api.post(
+        `/api/v1/boards/${boardId}/columns`,
+        columns
+      );
+      return response.data;
+    } catch (err) {
+      console.error(err);
+    }
+  },
+  deleteBoard: async (boardId) => {
+    try {
+      const response = await api.delete(`/api/v1/boards/${boardId}`);
+      return response.data;
+    } catch (err) {
+      console.error(err);
+    }
+  },
+};
+
+export const taskService = {
   createTask: async () => {
     try {
     } catch (err) {
